ENH: moved f16 functions to a submodule
