+++ src/FlightMechanicsSimulator.jl
@@ -1,11 +1,8 @@
 module FlightMechanicsSimulator
 
+export GD, R0, RAD2DEG, DEG2RAD
+
 include("constants.jl")
-include("adc.jl")
-include("engine.jl")
-include("aero.jl")
-include("sixdof.jl")
-include("rk4.jl")
-include("trimmer.jl")
+include("f16/F16.jl")  # F16 submodule
 
 end
