+++ src/f16/F16.jl
@@ -1,0 +1,13 @@
+module F16
+
+using ..FlightMechanicsSimulator
+
+include("constants.jl")
+include("adc.jl")
+include("engine.jl")
+include("aero.jl")
+include("sixdof.jl")
+include("rk4.jl")
+include("trimmer.jl")
+
+end
