+++ src/f16/adc.jl
