+++ test/aero.jl
@@ -15,52 +15,52 @@ using FlightMechanicsSimulator
 
     df = DataFrame!(CSV.File("data/damp.csv"))
     for case in eachrow(df)
-        rv1 = FlightMechanicsSimulator.damp(case.alpha)
+        rv1 = FlightMechanicsSimulator.F16.damp(case.alpha)
         @test isapprox(rv1, Array(case[2:end]), atol = 1.0e-14)
     end
 
     df = DataFrame!(CSV.File("data/cx.csv"))
     for case in eachrow(df)
-        rv1 = FlightMechanicsSimulator.CX(case.alpha, case.de)
+        rv1 = FlightMechanicsSimulator.F16.CX(case.alpha, case.de)
         @test isapprox(rv1, case.cx, atol = 1.0e-15)
     end
 
     df = DataFrame!(CSV.File("data/cy.csv"))
     for case in eachrow(df)
-        rv1 = FlightMechanicsSimulator.CY(case.beta, case.da, case.dr)
+        rv1 = FlightMechanicsSimulator.F16.CY(case.beta, case.da, case.dr)
         @test isapprox(rv1, case.cy, atol = 1.0e-15)
     end
 
     df = DataFrame!(CSV.File("data/cz.csv"))
     for case in eachrow(df)
-        rv1 = FlightMechanicsSimulator.CZ(case.alpha, case.beta, case.de)
+        rv1 = FlightMechanicsSimulator.F16.CZ(case.alpha, case.beta, case.de)
         @test isapprox(rv1, case.cz, atol = 1.0e-15)
     end
 
     df = DataFrame!(CSV.File("data/cm.csv"))
     for case in eachrow(df)
-        rv1 = FlightMechanicsSimulator.CM(case.alpha, case.de)
+        rv1 = FlightMechanicsSimulator.F16.CM(case.alpha, case.de)
         @test isapprox(rv1, case.cm, atol = 1.0e-15)
     end
 
     df = DataFrame!(CSV.File("data/aero_coeffs.csv"))
     for case in eachrow(df)
-        rv1 = FlightMechanicsSimulator.CL(case.alpha, case.beta)
+        rv1 = FlightMechanicsSimulator.F16.CL(case.alpha, case.beta)
         @test isapprox(rv1, case.cl, atol = 1.0e-15)
 
-        rv1 = FlightMechanicsSimulator.CN(case.alpha, case.beta)
+        rv1 = FlightMechanicsSimulator.F16.CN(case.alpha, case.beta)
         @test isapprox(rv1, case.cn, atol = 1.0e-15)
 
-        rv1 = FlightMechanicsSimulator.DLDA(case.alpha, case.beta)
+        rv1 = FlightMechanicsSimulator.F16.DLDA(case.alpha, case.beta)
         @test isapprox(rv1, case.dlda, atol = 1.0e-15)
 
-        rv1 = FlightMechanicsSimulator.DLDR(case.alpha, case.beta)
+        rv1 = FlightMechanicsSimulator.F16.DLDR(case.alpha, case.beta)
         @test isapprox(rv1, case.dldr, atol = 1.0e-15)
 
-        rv1 = FlightMechanicsSimulator.DNDA(case.alpha, case.beta)
+        rv1 = FlightMechanicsSimulator.F16.DNDA(case.alpha, case.beta)
         @test isapprox(rv1, case.dnda, atol = 1.0e-15)
 
-        rv1 = FlightMechanicsSimulator.DNDR(case.alpha, case.beta)
+        rv1 = FlightMechanicsSimulator.F16.DNDR(case.alpha, case.beta)
         @test isapprox(rv1, case.dndr, atol = 1.0e-15)
     end
 end
